[PATRO 23-7] move keycloak settings and user module to api project, add summary
