+++ src/api/app/Program.cs
@@ -1,7 +1,6 @@
+using Intive.Patronage2023.Api;
 using Intive.Patronage2023.Api.Configuration;
 using Intive.Patronage2023.Modules.Example.Api;
-using Intive.Patronage2023.Modules.Example.Application.User;
-using Intive.Patronage2023.Modules.Example.Domain;
 using Intive.Patronage2023.Shared.Abstractions;
 using Intive.Patronage2023.Shared.Abstractions.Commands;
 using Intive.Patronage2023.Shared.Abstractions.Queries;
