--- src/modules/example/domain/ApiKeycloakCredentials.cs
@@ -1,12 +1,0 @@
-namespace Intive.Patronage2023.Modules.Example.Domain;
-
-/// <summary>
-/// ApiKeycloakCredentials.
-/// </summary>
-public class ApiKeycloakCredentials
-{
-	/// <summary>
-	/// Secret.
-	/// </summary>
-	public string? Secret { get; set; }
-}
